+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Outline tega types.
 - Specify `TEGA_Solidity` interface.
 - Implement `TEGA_Solidity` interface.
+- Add `TEGA_Type` for type erasure.
 
 ## [0.0.0] - 2024-08-18